Refactor: move file, directory
data: local에서만 관련된 model
network: 통신 관련 파일
network/request, network/response: 실질 네트워크 통신에 사용된 body의 data class
network/service: 네트워크 통신 인터페이스
network/repository: 데이터 로직 처리 리파지토리
