+++ app/src/main/java/com/sgs/devcamp2/flametalk_android/data/dummy/FriendDummyData.kt
@@ -1,6 +1,6 @@
-package com.sgs.devcamp2.flametalk_android.domain.dummy
+package com.sgs.devcamp2.flametalk_android.data.dummy
 
-import com.sgs.devcamp2.flametalk_android.domain.model.response.friend.ProfilePreview
+import com.sgs.devcamp2.flametalk_android.network.response.friend.ProfilePreview
 
 /**
  * @author 박소연
