--- app/src/main/java/com/sgs/devcamp2/flametalk_android/domain/repository/chatlist/ChatListDao.kt
@@ -1,3 +1,0 @@
-package com.sgs.devcamp2.flametalk_android.domain.repository.chatlist
-
-interface ChatListDao
