--- app/src/main/java/com/sgs/devcamp2/flametalk_android/domain/repository/chatting/Chatting.kt
@@ -1,3 +1,0 @@
-package com.sgs.devcamp2.flametalk_android.domain.repository.chatting
-
-interface Chatting
