--- app/src/main/java/com/sgs/devcamp2/flametalk_android/domain/repository/user/UserDao.kt
@@ -1,4 +1,0 @@
-package com.sgs.devcamp2.flametalk_android.domain.repository.user
-
-interface UserDao {
-}
