+++ app/src/main/java/com/sgs/devcamp2/flametalk_android/network/NetworkModule.kt
@@ -1,7 +1,7 @@
 package com.sgs.devcamp2.flametalk_android.network
 
 import com.sgs.devcamp2.flametalk_android.data.repository.UserRepository
-import com.sgs.devcamp2.flametalk_android.service.UserService
+import com.sgs.devcamp2.flametalk_android.network.service.UserService
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
