+++ app/src/main/java/com/sgs/devcamp2/flametalk_android/network/repository/chatlist/ChatListDao.kt
@@ -1,0 +1,3 @@
+package com.sgs.devcamp2.flametalk_android.network.repository.chatlist
+
+interface ChatListDao
