+++ app/src/main/java/com/sgs/devcamp2/flametalk_android/network/repository/chatting/Chatting.kt
@@ -1,0 +1,3 @@
+package com.sgs.devcamp2.flametalk_android.network.repository.chatting
+
+interface Chatting
