+++ app/src/main/java/com/sgs/devcamp2/flametalk_android/network/repository/user/UserDao.kt
@@ -1,0 +1,4 @@
+package com.sgs.devcamp2.flametalk_android.network.repository.user
+
+interface UserDao {
+}
