+++ app/src/main/java/com/sgs/devcamp2/flametalk_android/network/response/friend/ProfilePreview.kt
@@ -1,4 +1,4 @@
-package com.sgs.devcamp2.flametalk_android.domain.model.response.friend
+package com.sgs.devcamp2.flametalk_android.network.response.friend
 
 data class ProfilePreview(
     val nickname: String,
