--- app/src/main/java/com/sgs/devcamp2/flametalk_android/service/ChatService.kt
@@ -1,4 +1,0 @@
-package com.sgs.devcamp2.flametalk_android.service
-
-class ChatService {
-}
